requirements.txt
speechpy
pytorch
